GS-63 refactor validator class
