+++ client/src/components/Animator/GameResults.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+interface GameResultsProps {
+
+}
+
+export default function AgeTransition (props: GameResultsProps) {
+    
+    return (
+        <div className="game-results">
+            
+        </div>
+    )
+}
